Updated File Structure
Now I think that the structure of each note folder is very complete.

There is

Note/
 |
  --> markdown/
  --> ipynb/
  --> data/
  --> includes

I should simply make this structure by default.  I could then populate
it with a note.md file in the markdown folder that is already setup to
track a note.ipynb!  This is super cool!